:recycle: Separate file for Colk datasource, delegate, helper
